Get rid of circular deps in Operation
