+++ src/types/Operation.ts
@@ -1,7 +1,8 @@
-import { MichelsonV1Expression } from "@taquito/rpc";
+import { MichelsonV1Expression, TransactionOperationParameter } from "@taquito/rpc";
+import { MANAGER_LAMBDA } from "@taquito/taquito";
+import { isEqual } from "lodash";
 
 import { Address, ContractAddress, ImplicitAddress } from "./Address";
-import { makeBatchLambda } from "../multisig/multisigUtils";
 import { ApproveOrExecute } from "../utils/tezos/types";
 
 export type TezTransfer = {
@@ -59,6 +60,11 @@ export type Operation =
   | ContractOrigination
   | ContractCall;
 
+const LAMBDA_HEADER: MichelsonV1Expression[] = [
+  { prim: "DROP" },
+  { prim: "NIL", args: [{ prim: "operation" }] },
+];
+
 export const makeMultisigApproveOrExecuteOperation = (
   contract: ContractAddress,
   entrypoint: ApproveOrExecute,
@@ -74,7 +80,7 @@ export const makeMultisigProposeOperation = (
   contract: ContractAddress,
   proposedOperations: Operation[]
 ): ContractCall => {
-  const lambdaActions = makeBatchLambda(proposedOperations);
+  const lambdaActions = toBatchLambda(proposedOperations);
   return makeContractCallOperation(contract, "propose", lambdaActions);
 };
 
@@ -92,3 +98,217 @@ const makeContractCallOperation = (
     amount,
   };
 };
+
+export const headlessLambda = (lambda: MichelsonV1Expression[]): MichelsonV1Expression[] => {
+  if (isEqual(lambda.slice(0, 2), LAMBDA_HEADER)) {
+    return lambda.slice(2);
+  }
+  return lambda;
+};
+
+export const toLambda = (operation: Operation): MichelsonV1Expression[] => {
+  switch (operation.type) {
+    case "tez":
+      switch (operation.recipient.type) {
+        case "implicit":
+          return MANAGER_LAMBDA.transferImplicit(operation.recipient.pkh, Number(operation.amount));
+        case "contract":
+          return MANAGER_LAMBDA.transferToContract(
+            operation.recipient.pkh,
+            Number(operation.amount)
+          );
+      }
+    // eslint-disable-next-line no-fallthrough
+    case "fa1.2":
+      return contractLambda(
+        operation,
+        FA12_TRANSFER_ARG_TYPES,
+        makeFA12TransactionParameter(operation)
+      );
+    case "fa2": {
+      return contractLambda(
+        operation,
+        FA2_TRANSFER_ARG_TYPES,
+        makeFA2TransactionParameter(operation)
+      );
+    }
+    case "delegation":
+      return MANAGER_LAMBDA.setDelegate(operation.recipient.pkh);
+    case "undelegation":
+      return MANAGER_LAMBDA.removeDelegate();
+    case "contract_origination":
+    case "contract_call":
+      throw new Error(`${operation.type} is not supported yet`);
+  }
+};
+
+/**
+ *
+ * @param operations - List of JSON operations
+ * @param network - Network is needed for fetching contract parameter elements in lambda
+ * @returns Lambda in MichelsonJSON (=Micheline) format
+ */
+
+export const toBatchLambda = (operations: Operation[]) => {
+  const opsLambdas = operations.map(operation => toLambda(operation)).flatMap(headlessLambda);
+
+  return [...LAMBDA_HEADER, ...opsLambdas];
+};
+
+export const FA2_TRANSFER_ARG_TYPES: MichelsonV1Expression = {
+  args: [
+    {
+      args: [
+        {
+          annots: ["%from_"],
+          prim: "address",
+        },
+        {
+          annots: ["%txs"],
+          args: [
+            {
+              args: [
+                {
+                  annots: ["%to_"],
+                  prim: "address",
+                },
+                {
+                  args: [
+                    {
+                      annots: ["%token_id"],
+                      prim: "nat",
+                    },
+                    {
+                      annots: ["%amount"],
+                      prim: "nat",
+                    },
+                  ],
+                  prim: "pair",
+                },
+              ],
+              prim: "pair",
+            },
+          ],
+          prim: "list",
+        },
+      ],
+      prim: "pair",
+    },
+  ],
+  prim: "list",
+};
+
+export const FA12_TRANSFER_ARG_TYPES: MichelsonV1Expression = {
+  args: [
+    {
+      annots: [":from"],
+      prim: "address",
+    },
+    {
+      args: [
+        {
+          annots: [":to"],
+          prim: "address",
+        },
+        {
+          annots: [":value"],
+          prim: "nat",
+        },
+      ],
+      prim: "pair",
+    },
+  ],
+  prim: "pair",
+};
+
+const contractLambda = (
+  operation: FA12Transfer | FA2Transfer,
+  argTypes: MichelsonV1Expression,
+  transactionParameter: TransactionOperationParameter
+) => {
+  return [
+    ...LAMBDA_HEADER,
+    {
+      prim: "PUSH",
+      args: [
+        { prim: "address" },
+        { string: operation.contract.pkh + "%" + transactionParameter.entrypoint },
+      ],
+    },
+    {
+      prim: "CONTRACT",
+      args: [argTypes],
+    },
+    // If contract is not valid then fail and rollback the whole transaction
+    [{ prim: "IF_NONE", args: [[{ prim: "UNIT" }, { prim: "FAILWITH" }], []] }],
+    { prim: "PUSH", args: [{ prim: "mutez" }, { int: "0" }] },
+    { prim: "PUSH", args: [argTypes, transactionParameter.value] },
+    { prim: "TRANSFER_TOKENS" },
+    { prim: "CONS" },
+  ];
+};
+export const makeFA12TransactionParameter = ({
+  sender,
+  recipient,
+  amount,
+}: FA12Transfer): TransactionOperationParameter => ({
+  entrypoint: "transfer",
+  value: {
+    prim: "Pair",
+    args: [
+      {
+        string: sender.pkh,
+      },
+      {
+        prim: "Pair",
+        args: [
+          {
+            string: recipient.pkh,
+          },
+          {
+            int: amount,
+          },
+        ],
+      },
+    ],
+  },
+});
+export const makeFA2TransactionParameter = ({
+  sender,
+  recipient,
+  tokenId,
+  amount,
+}: FA2Transfer): TransactionOperationParameter => ({
+  entrypoint: "transfer",
+  value: [
+    {
+      prim: "Pair",
+      args: [
+        {
+          string: sender.pkh,
+        },
+        [
+          {
+            prim: "Pair",
+            args: [
+              {
+                string: recipient.pkh,
+              },
+              {
+                prim: "Pair",
+                args: [
+                  {
+                    int: tokenId,
+                  },
+                  {
+                    int: amount,
+                  },
+                ],
+              },
+            ],
+          },
+        ],
+      ],
+    },
+  ],
+});
