+++ src/utils/tezos/helpers.test.ts
@@ -4,8 +4,6 @@ import axios from "axios";
 import {
   addressExists,
   getPkAndPkhFromSk,
-  makeFA12TransactionParameter,
-  makeFA2TransactionParameter,
   operationToTaquitoOperation,
   operationsToBatchParams,
 } from "./helpers";
@@ -22,12 +20,7 @@ import {
 } from "../../mocks/factories";
 import { makeAccountOperations } from "../../types/AccountOperations";
 import { MAINNET } from "../../types/Network";
-import {
-  ContractCall,
-  ContractOrigination,
-  FA12Transfer,
-  FA2Transfer,
-} from "../../types/Operation";
+import { ContractCall, ContractOrigination } from "../../types/Operation";
 jest.mock("@taquito/signer");
 jest.mock("./fakeSigner");
 jest.mock("axios");
@@ -66,88 +59,6 @@ describe("tezos utils helpers", () => {
       `${MAINNET.tzktApiUrl}/v1/accounts/${mockImplicitAddress(0).pkh}`
     );
     expect(result).toEqual(false);
-  });
-
-  describe("token transfers", () => {
-    test("makeFA2TransactionParameter", () => {
-      const operation: FA2Transfer = {
-        type: "fa2",
-        amount: "10",
-        tokenId: "5",
-        recipient: mockImplicitAddress(0),
-        contract: mockContractAddress(1),
-        sender: mockContractAddress(2),
-      };
-
-      expect(makeFA2TransactionParameter(operation)).toEqual({
-        entrypoint: "transfer",
-        value: [
-          {
-            prim: "Pair",
-            args: [
-              {
-                string: mockContractAddress(2).pkh,
-              },
-              [
-                {
-                  prim: "Pair",
-                  args: [
-                    {
-                      string: mockImplicitAddress(0).pkh,
-                    },
-                    {
-                      prim: "Pair",
-                      args: [
-                        {
-                          int: "5",
-                        },
-                        {
-                          int: "10",
-                        },
-                      ],
-                    },
-                  ],
-                },
-              ],
-            ],
-          },
-        ],
-      });
-    });
-
-    test("makeFA12TransactionParameter", () => {
-      const operation: FA12Transfer = {
-        type: "fa1.2",
-        amount: "110",
-        recipient: mockImplicitAddress(1),
-        contract: mockContractAddress(2),
-        tokenId: "0",
-        sender: mockContractAddress(1),
-      };
-
-      expect(makeFA12TransactionParameter(operation)).toEqual({
-        entrypoint: "transfer",
-        value: {
-          prim: "Pair",
-          args: [
-            {
-              string: mockContractAddress(1).pkh,
-            },
-            {
-              prim: "Pair",
-              args: [
-                {
-                  string: mockImplicitAddress(1).pkh,
-                },
-                {
-                  int: "110",
-                },
-              ],
-            },
-          ],
-        },
-      });
-    });
   });
 
   describe("operationToTaquitoOperation", () => {
